pyproject.toml
[project]
name = "biodiversityasset"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pdfminer.six>=20221105",
    "pandas>=2.0.0",
    "transformers>=4.21.0",
    "torch>=1.12.0",
    "tokenizers>=0.13.0",
    "tqdm>=4.64.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]
